Nova branch criada para não confundir com branch q estava funcionando.

+++ loja/src/main/java/br/com/alura/loja/daoBD/CarrinhoDaoInterface.java
@@ -14,6 +14,6 @@ public interface CarrinhoDaoInterface {
 	
 	public List<Carrinho> listarCarrinhos();
 	
-	public Carrinho procurarPorCodigo(int id);
+	public Carrinho procurarPorCodigo(long id);
 
 }
